requirements.txt
groq
langchain
langchain-community
langchain-core
langgraph
faiss-cpu
sentence-transformers
duckduckgo-search
streamlit
python-dotenv
sympy
gtts